pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=1.2.17",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "numpy>=2.4.2",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
    "torch>=2.10.0",
]
